requirements.txt
openai-whisper
TTS
vosk
pyaudio
webrtcvad
noisereduce
numpy
scipy
transformers
torch
pytest
pytest-cov
